+++ src/GUI/PackageComponent.h
@@ -296,11 +296,7 @@ signals:
 private:
     void createIconLayout(const QUrl &iconUrl){
         iconButton = new QLabel(this);
-        if(package->isAvailableInDB()) {
-            iconButton->setFixedSize(QSize(PACKAGE_ICON_SIZE, PACKAGE_ICON_SIZE));
-        } else {
-            iconButton->setFixedWidth(PACKAGE_ICON_SIZE);
-        }
+        iconButton->setFixedSize(QSize(PACKAGE_ICON_SIZE, PACKAGE_ICON_SIZE));
         iconButton->setAlignment(Qt::AlignHCenter | Qt::AlignVCenter);
         iconButton->setStyleSheet(PACKAGE_ICON_STYLESHEET);
         iconLayout = new QHBoxLayout;