+++ src/GUI/viewitems/SystemUpdatablePackageListView.cpp
@@ -48,9 +48,10 @@ void SystemUpdatablePackageListView::getSystemUpdatablePackages(const QVector<Pa
                 otherPackageList.append(pkg);
             }
         }
-        auto otherApplicationTitle = new OtherApplicationsWidgetItem(this);
-        boxLayout->addWidget(otherApplicationTitle);
-        
+        if(otherPackageList.size()) {
+            auto otherApplicationTitle = new OtherApplicationsWidgetItem(this);
+            boxLayout->addWidget(otherApplicationTitle);
+        }
         for(auto &pkg:otherPackageList){
             auto packageWidget = new PackageListWidgetItem(pkg, true, this);
             boxLayout->addWidget(packageWidget);