+++ src/GUI/viewitems/UserUpdatablePackageListView.cpp
@@ -63,9 +63,10 @@ void UserUpdatablePackageListView::getUserUpdatablePackages(const QVector<Packag
                 otherPackageList.append(pkg);
             }
         }
-        auto otherApplicationTitle = new OtherApplicationsWidgetItem(this);
-        boxLayout->addWidget(otherApplicationTitle);
-        
+        if(otherPackageList.size()) {
+            auto otherApplicationTitle = new OtherApplicationsWidgetItem(this);
+            boxLayout->addWidget(otherApplicationTitle);
+        }
         for(auto &pkg:otherPackageList){
             auto packageWidget = new PackageListWidgetItem(pkg, true, this);
             boxLayout->addWidget(packageWidget);